loader: move endpoint-specific functionality out to a new endpoint.go
